+++ NEWS.md
@@ -1,5 +1,17 @@
 Changelog
 
+# 0.9.16
+3rd July 2021
+
+## New
+
+* new config_db table `Settings`
+  + `setting` `value` pairs (character)
+* new `self$dateformat` reactive
+  + date of formats to be used in GPstat! shiny GUI
+  + default value is `2021-01-17` (YYYY-mm-dd or %Y-%-m-%d)
+  + value read from config_db `Settings` table, with the `setting` value of `dateformat`
+
 # 0.9.15
 16th June 2021